% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_auth.R
\docType{data}
\name{auth}
\alias{auth}
\title{auth}
\format{An object of class \code{character} of length 2.}
\usage{
auth
}
\description{
A single author/inventor name ("Ventor Craig") for internal use in function tests
}
\keyword{datasets}

