% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_type_.R
\name{author_type_}
\alias{author_type_}
\title{single author url by type}
\usage{
author_type_(query, author_type = "NULL")
}
\arguments{
\item{query}{a single author name (e.g. Kirk James or Kirk J)}

\item{author_type}{default, crossref or pubmed}
}
\value{
a url
}
\description{
used internally to format urls for single author names for default, crossref or pubmed. Used in author_search
}
\examples{
\dontrun{lens_authors("Kirk James", author_type = "crossref")}
}

