% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_country.R
\docType{data}
\name{country}
\alias{country}
\title{country}
\format{An object of class \code{data.frame} with 273 rows and 15 columns.}
\usage{
country
}
\description{
A table of country names and country codes from the countrycode package. Note that the first column has been renamed. The country_name column is extracted for use in searches of the Lens.
}
\keyword{datasets}

