% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_drones_query.R
\docType{data}
\name{drones}
\alias{drones}
\title{drones}
\format{An object of class \code{character} of length 4.}
\usage{
drones
}
\description{
A character vector containing four search terms for drone technology (with the odd drone bee showing up in there). Used for testing larger queries and booleans.
drones  <- c("drone", "drones", "Unmanned Aerial Vehicle", "UAV")
}
\keyword{datasets}
