% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kenya_lake_species.R
\docType{data}
\name{kenya_lake_species}
\alias{kenya_lake_species}
\title{kenya_lake_species}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 379 rows and 1 columns.}
\usage{
kenya_lake_species
}
\description{
A raw list of species names linked to searches of the scientific literature (Web of Science core collection)
}
\keyword{datasets}
