% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_authors.R
\name{lens_authors}
\alias{lens_authors}
\title{Search using author names}
\usage{
lens_authors(author, author_boolean = "NULL", author_type = "NULL")
}
\arguments{
\item{author}{a single author name or vector of author names (quoted)}

\item{author_boolean}{OR or AND}

\item{author_type}{default, crossref or pubmed}
}
\value{
a url
}
\description{
search for author names in the text of patent documents using the Lens.
}
\examples{
\dontrun{lens_authors("Venter Craig")}
}

