% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_inventor.R
\name{lens_inventor}
\alias{lens_inventor}
\title{Search the Lens using inventor names}
\usage{
lens_inventor(inventor = "NULL", inventor_boolean = "NULL")
}
\arguments{
\item{inventor}{an inventor name (e.g. Kirk James) or vector of inventor names (quoted).}

\item{inventor_boolean}{Either AND or OR (default) for use with multiple inventor names (quoted).}
}
\value{
a url
}
\description{
Search the lens using one or more inventor names
}
\examples{
\dontrun{(lens_inventor("Venter Craig"))}
}

