% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_inventor_keyword.R
\name{lens_inventor_keyword}
\alias{lens_inventor_keyword}
\title{Search the Lens using inventor names and key terms.}
\usage{
lens_inventor_keyword(inventor = "NULL", inventor_boolean = "NULL",
  query = "NULL", type = "NULL", boolean = "NULL")
}
\arguments{
\item{inventor}{An inventor name or vector of inventor names.}

\item{inventor_boolean}{OR or AND}

\item{query}{A search term or set of search terms}

\item{type}{fulltext = default, "ti" = title, "ab" = abstract, "tac" = title or abstract or claims. See ops_urls for details.}

\item{boolean}{for vector of search terms OR or AND}
}
\value{
a url for use in ops_iterate
}
\description{
Use a single inventor name or vector of inventor names with one or more search terms. Inventor names should be in the format "Surname First Name".
}
\examples{
\dontrun{lens_inventor_keyword("Kerr Russell", query = "bahamas")}
}

