% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_inventors_keyword.R
\name{lens_inventors_keyword}
\alias{lens_inventors_keyword}
\title{Search the Lens using inventor names and key terms.}
\usage{
lens_inventors_keyword(inventor = "NULL", inventor_boolean = "NULL",
  query = "NULL", type = "NULL", boolean = "NULL")
}
\arguments{
\item{inventor}{An inventor name or vector of inventor names.}

\item{inventor_boolean}{OR or AND}

\item{query}{A search term or vector of search terms}

\item{type}{fulltext = default, "ti" = title, "ab" = abstract, "tac" = title or abstract or claims. See ops_urls for details.}

\item{boolean}{for vector of search terms OR or AND}
}
\value{
a url for use in ops_iterate
}
\description{
Use a single inventor name or vector of inventor names with one or more search terms. Inventor names should be in the format "Surname First Name".
}
\examples{
\dontrun{lens_inventors_keyword("Kerr Russell", query = "bahamas")}
\dontrun{lens_inventors_keyword(inventor = "Pomponi Shirley", query = "bahamas")}
\dontrun{lens_inventors_keyword(inventor = auth, inventor_boolean = "OR", query = synbio, type = "tac", boolean = "OR")}
\dontrun{lens_inventors_keyword(inventor = "Venter Craig", query = c("synthetic genomes", "synthetic genomics"), type = "tac", boolean = "OR")}
\dontrun{lens_inventors_keyword(inventor = "Venter Craig", query = country$country_name, boolean = "OR")}
}
