% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_parse.R
\name{lens_parse}
\alias{lens_parse}
\title{Parse data from lens_search to a data.frame or tibble}
\usage{
lens_parse(data)
}
\arguments{
\item{data}{html from the lens retrieved using rvest read_html}
}
\value{
a tibble.
}
\description{
Used internally to parse the results from lens_search.
}
\note{
By default Lens returns publication numbers as e.g. "WO 2015/199614 A1"
  with spaces between the country code and the kind code. This format is not
  used by any other database and is not used by the Lens itself for its new upload
  service where the format is WO_2015/199614_A1. To promote interoperability
  between databases Lens parse removes the spaces and also the "/" in Lens
  numbers. Note that the impact of the removal of "/" on retrieval from other
  databases requires exploration. In order to permit the upload of patent
  numbers back into the Lens the spaces in the Lens publication numbers are
  replaced with "_" as in "WO_2015/199614_A1".
}
\examples{
\dontrun{df <- lens_urls("drones")}
\dontrun{data <- lens_urls(synbio, boolean = "OR") \%>\% lens_parse()}
}
