% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_query_.R
\name{lens_query_}
\alias{lens_query_}
\title{Search using keyword combinations in clauses (internal)
Used internally}
\usage{
lens_query_(query, boolean, q1, stemming = FALSE, families = "NULL")
}
\arguments{
\item{query}{a search term or set of search terms}

\item{boolean}{AND/OR}

\item{q1}{a second set of search terms}

\item{stemming}{Whether to stem the terms, default is false}

\item{families}{Whether to restrict the search to families only.}
}
\value{
a url
}
\description{
Search using keyword combinations in clauses (internal)
Used internally
}
\examples{
\dontrun{
# search for terms using AND
lens_query_("kenya", "AND", "crater lake"),
# restrict results to families
lens_query_("kenya", "AND", "crater lake", families = TRUE)}
}

