% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_tidy_names.R
\name{lens_tidy_names}
\alias{lens_tidy_names}
\title{Tidy column names for a dataset imported from the Lens}
\usage{
lens_tidy_names(df)
}
\arguments{
\item{df}{A data frame imported from a download of a Lens Collection.}
}
\value{
a data.frame
}
\description{
Converts column names to lower case, spaces to underscores and removes brackets from datasets downloaded from the Lens patent database.
}
\examples{
\dontrun{lens_tidy_names(synbio, boolean = "OR", type = "title", rank_family = TRUE)}
}
