% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_urls_.R, R/zzz.R
\name{lens_urls_}
\alias{lens_urls_}
\title{Internal version of lens_urls}
\usage{
lens_urls_(query, boolean = "NULL", type = "NULL", applicant = NULL,
  applicant_boolean = "NULL", inventor = NULL, inventor_boolean = "NULL",
  publn_date_start = NULL, publn_date_end = NULL,
  filing_date_start = NULL, filing_date_end = NULL, rank_family = "NULL",
  rank_citing = "NULL", rank_sequences = "NULL",
  rank_latest_publn = "NULL", rank_earliest_publn = "NULL",
  rank_latest_filing = "NULL", rank_earliest_filing = "NULL",
  jurisdiction = "NULL", families = "NULL", timer = 20, results = NULL,
  stemming = "NULL")

lens_urls_(query, boolean = "NULL", type = "NULL", applicant = NULL,
  applicant_boolean = "NULL", inventor = NULL, inventor_boolean = "NULL",
  publn_date_start = NULL, publn_date_end = NULL,
  filing_date_start = NULL, filing_date_end = NULL, rank_family = "NULL",
  rank_citing = "NULL", rank_sequences = "NULL",
  rank_latest_publn = "NULL", rank_earliest_publn = "NULL",
  rank_latest_filing = "NULL", rank_earliest_filing = "NULL",
  jurisdiction = "NULL", families = "NULL", timer = 20, results = NULL,
  stemming = "NULL")
}
\arguments{
\item{query}{A search string or vector of search terms (quoted)}

\item{boolean}{Select the type of boolean ("OR" or "AND") where using multiple search terms.}

\item{type}{Either fulltext (default), title, abstract, claims, or "tac" for 'title or abstract or claims' (quoted).}

\item{applicant}{An applicant name or vector of applicant names}

\item{applicant_boolean}{"AND" or "OR".}

\item{inventor}{An inventor name or vector of inventor names.}

\item{inventor_boolean}{"AND" or "OR".}

\item{publn_date_start}{Publication date limit to start at as YYYMMDD (numeric).}

\item{publn_date_end}{Publication date limit to end at as YYYMMDD (numeric).}

\item{filing_date_start}{Filing date limit to start at as YYYMMDD (numeric).}

\item{filing_date_end}{Filing date limit to end at as YYYMMDD (numeric).}

\item{rank_family}{Whether to sort the Lens results by the number of family
members (descending). Useful for retrieving important documents. See
details.}

\item{rank_citing}{Whether to sort the Lens results by the top citing
(descending). Useful for retrieving important documents. See details.}

\item{rank_sequences}{Rank results on whether the documents contain a dna or amino
acid sequence. See details.
#' @param rank_earliest_publn Rank the results by the earliest publication date (earliest publshed).}

\item{rank_latest_publn}{Rank the results by the latest publication date (most recently published).}

\item{rank_earliest_publn}{Rank the results by the latest publication date (most recently published).}

\item{rank_latest_filing}{Rank the results by the latest filing date.}

\item{rank_earliest_filing}{Rank the results by the earliest filing date.}

\item{jurisdiction}{Limit the search to a single jurisdiction (default is
all) e.g. "US" or choose inbuilt group "main" for the United States (US),
European Patent Office (EP), Japan (JP) or the World Intellectual Property
Organization (WO) for the Patent Cooperation Treaty.}

\item{families}{Either return the publication count and family numbers or if
TRUE (default) return the patent families (deduplicates a set of
publications to the first publication of the root "priority" or first
filing).}

\item{timer}{Where retrieving over 50 results, the delay between sending requests to the Lens (default is 20 seconds, used internally by ops_iterate()).}

\item{results}{The number of results to return, either 50 or 500 (maximum).}

\item{stemming}{Word stemming is set to FALSE by default.}
}
\value{
a data.frame or tibble
}
\description{
This function builds urls to search the Lens patent database. It is used internally in the lens_search() function. The default search groups documents by family and will return up to 50 results per page. The maximum number of results that can be retrieved is 500 (10 pages). For larger results sets use the free Lens online Collection facility to download upto 10,000 records. See details for information on the use of ranking and date measures to sort the data.

members (descending). Useful for retrieving important documents. See
  details.
  acid sequence. See details.
  #' @param rank_earliest_publn Rank the results by the earliest publication date (earliest publshed).
  all) e.g. "US" or choose inbuilt group "main" for the United States (US),
  European Patent Office (EP), Japan (JP) or the World Intellectual Property
  Organization (WO) for the Patent Cooperation Treaty.
  TRUE (default) return the patent families (deduplicates a set of
  publications to the first publication of the root "priority" or first
  filing).
  possible to rank by family scores but not family scores and latest
  publications or earliest publications. The suggested work flow is to
  retrieve the latest publications, then rank by family and then rank_citing.
  This will allow the most recent and the most important documents to be
  retrieved in three steps for a given query.
}
\details{
Only one ranking measure may be used per query. For example, it is
  possible to rank by family scores but not family scores and latest
  publications or earliest publications. The suggested work flow is to
  retrieve the latest publications, then rank by family and then rank_citing.
  This will allow the most recent and the most important documents to be
  retrieved in three steps for a given query.
}
\note{
The default connector between fields e.g. key terms and applicants and inventors is "AND".
}
\examples{
\dontrun{lens_urls("synthetic biology")}
}

