% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_one_url.R
\docType{data}
\name{one_url}
\alias{one_url}
\title{one_url}
\format{An object of class \code{character} of length 1.}
\usage{
one_url
}
\description{
A single url for use in testing.
}
\keyword{datasets}

