% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_synbio_query.R
\docType{data}
\name{synbio}
\alias{synbio}
\title{synbio}
\format{An object of class \code{character} of length 7.}
\usage{
synbio
}
\description{
A character vector containing a range of search terms for synthetic biology. Used for testing larger queries and booleans.
synbio <- c("synthetic biology", "synthetic genomics", "synthetic genome", "synthetic genomes", "biological parts", "genetic circuit", "genetic circuits")
}
\keyword{datasets}

