% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_applicants.R
\docType{data}
\name{three_applicants}
\alias{three_applicants}
\title{three_applicants}
\format{An object of class \code{character} of length 3.}
\usage{
three_applicants
}
\description{
A vector with three applicant names for use in testing.
}
\keyword{datasets}

