% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_three_authors.R
\docType{data}
\name{three_authors}
\alias{three_authors}
\title{three_authors}
\format{An object of class \code{character} of length 3.}
\usage{
three_authors
}
\description{
A vector with three author/inventor names for use in testing.
}
\keyword{datasets}

