% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_three_urls.R
\docType{data}
\name{three_urls}
\alias{three_urls}
\title{three_urls}
\format{An object of class \code{character} of length 3.}
\usage{
three_urls
}
\description{
A set of three urls for test searching the Lens patent database
[1] "https://www.lens.org/lens/search?p=0&q=%22synthetic+biology%22&n=50&f=true"
[2] "https://www.lens.org/lens/search?p=1&q=%22synthetic+biology%22&n=50&f=true"
[3] "https://www.lens.org/lens/search?p=2&q=%22synthetic+biology%22&n=50&f=true"
}
\keyword{datasets}

