% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_applicants.R
\name{lens_applicants}
\alias{lens_applicants}
\title{Search patent applicant names in the Lens}
\usage{
lens_applicants(applicant = NULL, applicant_boolean = "NULL")
}
\arguments{
\item{applicant}{an applicant name or vector of applicant names (character)}

\item{applicant_boolean}{OR or AND operators}
}
\value{
a URL
}
\description{
Search for patent applicant names in the Lens patent database.
}
\examples{
\dontrun{lens_applicant("Synthetic Genomics")}
}

