% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_ipc.R
\name{lens_ipcs}
\alias{lens_ipcs}
\title{Search the Lens using International Patent Classification (IPC) codes}
\usage{
lens_ipcs(ipc, ipc_boolean = "NULL")
}
\arguments{
\item{ipc}{An ipc code or vector of codes (character)}

\item{ipc_boolean}{"AND" or "OR" (quoted)}
}
\value{
a url
}
\description{
Search the Lens using individual or combinations of IPC Codes.
}
\details{
IPC codes on the group level must include a forward slash e.g.
  A61K31/00 or C12N15/82 or the Lens will not display results. A forward
  slash is not necessary on the subclass level e.g. A61K or C12N.
}
\examples{
\dontrun{lens_ipc(A61K31/00)}
}

