% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lense_parse.R
\name{lens_parse}
\alias{lens_parse}
\title{Parse data from lens_search to a data.frame or tibble}
\usage{
lens_parse(data)
}
\arguments{
\item{data}{html from the lens retrieved using rvest read_html}
}
\value{
a tibble.
}
\description{
Used internally to parse the results from lens_search.
}
\examples{
\dontrun{data <- lens_urls(synbio, boolean = "OR") \%>\% lens_parse()}
}

