% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_search.R
\name{lens_search}
\alias{lens_search}
\title{Search the Lens Patent Database}
\usage{
lens_search(query, boolean = "NULL", type = "NULL", pub_date_start = NULL,
  pub_date_end = NULL, filing_date_start = NULL, filing_date_end = NULL,
  rank_family = "NULL", rank_citing = "NULL", rank_sequences = "NULL",
  latest_publication = "NULL", earliest_publication = "NULL",
  latest_filing = "NULL", earliest_filing = "NULL", families = "TRUE",
  results = NULL, timer = 30, inventor = NULL,
  inventor_boolean = "NULL")
}
\arguments{
\item{query}{One or more terms to search in the patent database.}

\item{boolean}{Select the type of boolean ("OR" or "AND") where using multiple search terms.}

\item{type}{Either fulltext (default), title, abstract, claims, or "tac" for 'title or abstract or claims'.
abstract and claims (tac). Quoted.}

\item{pub_date_start}{Publication date limit to start at as YYYMMDD (numeric).}

\item{pub_date_end}{Publication date limit to end at as YYYMMDD (numeric).}

\item{filing_date_start}{Filing date limit to start at as YYYMMDD (numeric).}

\item{filing_date_end}{Filing date limit to end at as YYYMMDD (numeric).}

\item{rank_family}{Whether to sort the Lens results by the number of family
members (descending). Useful for retrieving important documents. See
details.}

\item{rank_citing}{Whether to sort the Lens results by the top citing
(descending). Useful for retrieving important documents. See details.}

\item{rank_sequences}{Rank results on whether the documents contain a dna or amino
acid sequence. See details.}

\item{latest_publication}{Sort results by the latest publication date.
Useful for retrieving the most recent documents. See details.}

\item{earliest_publication}{Sort results by the earliest publication date.
See details.}

\item{latest_filing}{Sort results by the latest filing (priority) date.
Useful for identifying the latest filings (note that they are the latest
filings that have a publication). See details.}

\item{earliest_filing}{Sort results by the earliest publication date. See
details.}

\item{families}{Either return the publication count and family numbers or if
TRUE (default) return the patent families (deduplicates a set of
publications to the first publication of the root "priority" or first
filing).}

\item{results}{The number of results to return, either 50 or 500 (maximum).}

\item{timer}{Where retrieving over 50 results, the delay between sending requests to the Lens (default is 30 seconds, used internally by ops_iterate()).}

\item{inventor}{An inventor name or vector of inventor names. Use the convention surname and first name (family name and given name) for best results.}

\item{inventor_boolean}{Either "OR" or "AND"}
}
\value{
a data.frame or tibble
}
\description{
This function allows for the construction of complex queries to search and retrieve data from the Lens. The default search groups documents by family and will return up to 50 results per page. The maximum number of results that can be retrieved is 500 (10 pages). For larger results sets use the free Lens online Collection facility to download upto 10,000 records. See details for information on the use of ranking and date measures to sort the data.
}
\details{
Only one ranking measure may be used per query. For example, it is
  possible to rank by family scores but not family scores and latest
  publications or earliest publications. The suggested work flow is to
  retrieve the latest publications, then rank by family and then rank_citing.
  This will allow the most recent and the most important documents to be
  retrieved in three steps for a given query. In patent documents the convention is to list the surname (family name) and then the first names (given names). Name reversals can and do occur but best practice is to use e.g "Kirk James T" rather than "James T Kirk".
}
\examples{
\dontrun{lens_search("synthetic biology", timer = 30)}
}

